Fix randomization algorithm to shuffle all photos.
Previously, for each "talk", we were picking 5 photos at random
and using those. After recent playtesting, it turned out that
we experienced a high number of repeated photos, and several
photos that never got picked.

Instead, what we do now is to shuffle all the photos as if they
were a deck of cards, and "deal" five photos off the top to form
a talk. We reshuffle when we run out of photos.